Refactored and sorted in packages, removed base files (Hello ..etc)
