--- src/main/java/org/example/Hello.java
@@ -1,14 +1,0 @@
-package org.example;
-
-public class Hello {
-
-    private String name;
-
-    public Hello(String name) {
-        this.name = name;
-    }
-
-    public String getHello(){
-        return name;
-    }
-}
