--- src/main/java/org/example/HelloApplication.java
@@ -1,9 +1,0 @@
-package org.example;
-
-import jakarta.ws.rs.core.Application;
-import jakarta.ws.rs.ApplicationPath;
-
-@ApplicationPath("api")
-public class HelloApplication extends Application {
-    // Needed to enable Jakarta REST and specify path.
-}
