+++ src/main/java/org/example/pets/exception/NotFoundExceptionMapper.java
@@ -1,0 +1,4 @@
+package org.example.pets.exception;
+
+public class NotFoundExceptionMapper {
+}
